services/citizen-vault/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests[socks]==2.31.0
cryptography==41.0.5
python-multipart==0.0.6
Pillow==10.1.0
python-dotenv==1.0.0
